+++ library/common/src/main/java/com/google/android/exoplayer2/DefaultControlDispatcher.java
@@ -21,19 +21,15 @@ import static java.lang.Math.min;
 /** Default {@link ControlDispatcher}. */
 public class DefaultControlDispatcher implements ControlDispatcher {
 
-  /** The default fast forward increment, in milliseconds. */
-  public static final int DEFAULT_FAST_FORWARD_MS = 15_000;
-  /** The default rewind increment, in milliseconds. */
-  public static final int DEFAULT_REWIND_MS = 5000;
-
-  private static final int MAX_POSITION_FOR_SEEK_TO_PREVIOUS = 3000;
-
   private final long rewindIncrementMs;
   private final long fastForwardIncrementMs;
+  private final boolean rewindAndFastForwardIncrementsSet;
 
   /** Creates an instance. */
   public DefaultControlDispatcher() {
-    this(DEFAULT_FAST_FORWARD_MS, DEFAULT_REWIND_MS);
+    fastForwardIncrementMs = C.TIME_UNSET;
+    rewindIncrementMs = C.TIME_UNSET;
+    rewindAndFastForwardIncrementsSet = false;
   }
 
   /**
@@ -47,6 +43,7 @@ public class DefaultControlDispatcher implements ControlDispatcher {
   public DefaultControlDispatcher(long fastForwardIncrementMs, long rewindIncrementMs) {
     this.fastForwardIncrementMs = fastForwardIncrementMs;
     this.rewindIncrementMs = rewindIncrementMs;
+    rewindAndFastForwardIncrementsSet = true;
   }
 
   @Override
@@ -69,39 +66,21 @@ public class DefaultControlDispatcher implements ControlDispatcher {
 
   @Override
   public boolean dispatchPrevious(Player player) {
-    Timeline timeline = player.getCurrentTimeline();
-    if (timeline.isEmpty() || player.isPlayingAd()) {
-      return true;
-    }
-    boolean isUnseekableLiveStream =
-        player.isCurrentWindowLive() && !player.isCurrentWindowSeekable();
-    if (player.hasPreviousWindow()
-        && (player.getCurrentPosition() <= MAX_POSITION_FOR_SEEK_TO_PREVIOUS
-            || isUnseekableLiveStream)) {
-      player.seekToPreviousWindow();
-    } else if (!isUnseekableLiveStream) {
-      player.seekTo(/* positionMs= */ 0);
-    }
+    player.seekToPrevious();
     return true;
   }
 
   @Override
   public boolean dispatchNext(Player player) {
-    Timeline timeline = player.getCurrentTimeline();
-    if (timeline.isEmpty() || player.isPlayingAd()) {
-      return true;
-    }
-    if (player.hasNextWindow()) {
-      player.seekToNextWindow();
-    } else if (player.isCurrentWindowLive() && player.isCurrentWindowDynamic()) {
-      player.seekToDefaultPosition();
-    }
+    player.seekToNext();
     return true;
   }
 
   @Override
   public boolean dispatchRewind(Player player) {
-    if (isRewindEnabled() && player.isCurrentWindowSeekable()) {
+    if (!rewindAndFastForwardIncrementsSet) {
+      player.seekBack();
+    } else if (isRewindEnabled() && player.isCurrentWindowSeekable()) {
       seekToOffset(player, -rewindIncrementMs);
     }
     return true;
@@ -109,7 +88,9 @@ public class DefaultControlDispatcher implements ControlDispatcher {
 
   @Override
   public boolean dispatchFastForward(Player player) {
-    if (isFastForwardEnabled() && player.isCurrentWindowSeekable()) {
+    if (!rewindAndFastForwardIncrementsSet) {
+      player.seekForward();
+    } else if (isFastForwardEnabled() && player.isCurrentWindowSeekable()) {
       seekToOffset(player, fastForwardIncrementMs);
     }
     return true;
@@ -142,22 +123,24 @@ public class DefaultControlDispatcher implements ControlDispatcher {
 
   @Override
   public boolean isRewindEnabled() {
-    return rewindIncrementMs > 0;
+    return !rewindAndFastForwardIncrementsSet || rewindIncrementMs > 0;
   }
 
   @Override
   public boolean isFastForwardEnabled() {
-    return fastForwardIncrementMs > 0;
+    return !rewindAndFastForwardIncrementsSet || fastForwardIncrementMs > 0;
   }
 
   /** Returns the rewind increment in milliseconds. */
-  public long getRewindIncrementMs() {
-    return rewindIncrementMs;
+  public long getRewindIncrementMs(Player player) {
+    return rewindAndFastForwardIncrementsSet ? rewindIncrementMs : player.getSeekBackIncrement();
   }
 
   /** Returns the fast forward increment in milliseconds. */
-  public long getFastForwardIncrementMs() {
-    return fastForwardIncrementMs;
+  public long getFastForwardIncrementMs(Player player) {
+    return rewindAndFastForwardIncrementsSet
+        ? fastForwardIncrementMs
+        : player.getSeekForwardIncrement();
   }
 
   // Internal methods.